add files for test, refactor
