+++ src/file/buses.txt
@@ -1,0 +1,6 @@
+59  Пазик 796332
+105 Пазик 311975
+205 BMW 350361
+471 BMW 944888
+477 ГАЗ 699224
+652 Пазик 437011
