+++ src/file/students.txt
@@ -1,0 +1,7 @@
+711 UW-45 4.03
+320 VP-27 4.25
+257 SD-17 1.0
+475 IHP-24 3.47
+797 XUT-36 4.41
+195 YBJ-9 2.32
+617 EZT-41 1.65
