+++ src/file/users
@@ -1,0 +1,5 @@
+Юра Ka4DGUZXL SNEHFcZH@google.org
+Женя 8aC4TM XPaBPPVD@inbox.by
+Коля b9DPF6MS ELIWNRC@google.by
+Юра J32SMF46NA NWJDcEV@mail.ru
+Вероника K8ATKYNYN SXaLJPOQ@inbox.org
