+++ src/main/java/client/SaveToFile.java
@@ -1,4 +1,4 @@
-package savetofile;
+package client;
 
 import java.io.File;
 import java.io.FileWriter;
